Update patch set 2
Patch Set 2:

Steve,
thanks for the pointer. Quarx is already working on it. One question though: Should we do it like for Wimax (with a LocationProxyHelper class in frameworks/base) or like the AOSP patch (which we would need to change a little, to use ServiceManager instead of only starting the service)? The AOSP patch approach + jar in BOOTCLASSPATH looks a little cleaner and more extendible to me.

Koushik,
thanks for the hint about Dex2Jar. Didn't know about it yet. If it only worked as advertised...it looks like it's in early development stages. At least the for loop resolutions it provides are plain wrong.

Patch-set: 2