Update patch set 2
Patch Set 2:

in your case you need to maintain the hole script. If you would do it with the stuff that is allready there you would only need to maintain your device specific stuff. 

here is how we do it on the Galaxy S device where we have a totally different updater-script than htc.

https://github.com/CyanogenMod/android_device_samsung_aries-common/blob/gingerbread/releasetools/aries_ota_from_target_files

in you case you would have to port any change that is done the the squisher.

Patch-set: 2